reviewboard: require client to advertise capability knowledge; r=smacleod
The previous commit introduced "client capabilities" and a mechanism
where a server can advertise minimum required capabilities and clients
can abort if they are older than what the server requires. But there is
a bootstrap problem: old clients don't know to look for the minimum
required capabilities and thus ignore them completely. This could lead
to trouble.

With this commit, the client advertises to the server that it knows how
to apply capability requirements. If the server doesn't see this
advertisement, it assumes the request is coming from an old client and
will reject the request, telling the client to upgrade. Deployment of
this commit to the server thus constitutes a flag day and will require
all deployed clients to be upgraded. After this commit is deployed, it
is safe to assume that any newly-advertised requirements will be
respected by the client. This enables us to more quickly break backwards
compatibility with old clients since clients can easily be marked as
incompatible.

--HG--
extra : commitid : JWqrGZOzQFe
extra : rebase_source : aa2b5e5c63e0ce7f9637892a85c120b48ffbab08