ansible/hgmo: produce bundles for more repositories
Bundles are being added for build/talos, esr38, b2g-inbound, and several
project branches that showed up in the top consumers list. As part of
this, we now more aggressively schedule bundle creation: every 30
minutes for large repositories. There may be some overlap in production,
but the server is beefy and it should be able to handle it.

--HG--
extra : commitid : Djt65rKFZDD