testing: don't adjust sys.path when invoking nose tests
As evidenced by the TODO, this code was suspect. Tests continue to pass
with this change.

The reason for this change is that nose tests were picking up the older
version of requests/urllib3 sitting in pylib/requests and hitting a bug
with modern versions of Python and PyOpenSSL. requests is available in
the virtualenv. Removing the addition of the pylib/ directories from the
module search path made the Selenium tests just work.

--HG--
extra : commitid : KbixKa91kGo