reviewboard: add sleep to patch over test failure (bug 1184602)
This is not the ideal workaround, but it appears to work, at least on my
machine. I have no clue what the race condition is and thus don't have
any leads on where to look to fix it. This seems like the easiest
solution to correcting a test failure.

--HG--
extra : commitid : 7eeWAXRwum2