mozreview: start passing cookie authentication when auto publishing (Bug 1184079). r=gps
We were not using provided cookie credentials to authenticate with
Review Board when automatically publishing review requests on push. We
now use these credentials if they are provided.

Also, we use the RBClient context manager from reviewboardmods.pushhooks
to avoid issues with the RBTools cache and cookie files that we hit with
the RBTools version we are using. Once we are able to upgrade to a
newer version of RBTools we should be able to remove the hacks here as
well.

--HG--
extra : commitid : 3hqxu3fYgvk
extra : rebase_source : 2c138a128ab2f1f9d8d9724d67f591de247a4569