ansible/hg-web: configure hg user's ssh settings (bug 1187090); r=fubar
This is a simple move of logic from docker-hg-web to hg-web.
docker-hg-web was based on the state of an actual hgweb server and the
Puppet configuration. So this should be trusted.

I confirmed as part of this refactoring that the ssh_config file
contents in Ansible match what is in production.

Of course, we'll eventually need to rewrite the ssh_config file to be a
template since the hostname of the hgssh server in the Docker cluster
will almost certainly not be "hg.mozilla.org" (unless we install a hosts
entry). But that's for another day.

--HG--
rename : ansible/roles/docker-hg-web/files/ssh-config => ansible/roles/hg-web/files/hg-ssh-config
extra : commitid : HXc4mClP3mG
extra : rebase_source : 66138b216b882b31aa1654e598a0ac961ebaa294
extra : histedit_source : 8876abc2fc3ecb7ab203445b605d0439550b522b