ansible/hg-ssh: use a single CRON entry for bundle generation (bug 1191128); r=fubar
We add a file containing the list of repositories whose bundles to
generate. It should be identical to the list of repositories in the CRON
jobs today.

We install a new, single CRON entry that generates bundles from this
file.

Note: Ansible won't delete the existing CRON entries. We'll want to
manually edit the crontab on hgssh1 when this is deployed.

--HG--
extra : commitid : 5EzIH5U2Gnz
extra : rebase_source : b6e5237bf9acbe095f3be24001c92927f21066cd