scripts: advertise CDN in bundle manifests (bug 1185261); r=fubar
We now have a CloudFront CDN for serving Mercurial bundles at
https://hg.cdn.mozilla.net/. Let's advertise the CDN URLs in the bundle
manifest.

Because of SNI concerns, we don't yet advertise the CDN URL as the
primary URL. Once we're satisfied that things won't break, we can bump
the CDN to the primary URL.

We introduce the ``cdn`` and ``requiresni`` attributes on the CDN
manifest entries to indicate the URL belongs to a CDN and our
CloudFront URL requires SNI. The use case for ``cdn`` isn't as
strong as ``requiresni`` (which is required for clients on old
Pythons to filter out URLs they won't be able to talk to), but it
can be useful for clients that wish to prefer connecting to a CDN
over other servers.

--HG--
extra : commitid : E5PkvL2p5Z7
extra : rebase_source : 9afc3b99f61a7e6d6f25b1347bb3f13540215019
extra : amend_source : c3cfc56559420383caf3921ccb433bdcf1bfae15