pushlog: change printed messages (bug 1197466); r=smacleod
The messages that the pushlog extension prints during push are formatted
inconsistently with the rest of Mercurial. They have capital letters and
periods where other messages do not. Personally, I like the well-formed
sentences. But this isn't what upstream does. So we change the
formatting.

The "Trying to insert" message isn't necessary because both the success
and failure cases print a message. The only time this would be useful is
if there is some kind of hang opening the pushlog database. We've had no
reports of this happening and I see little value in retaining this
message. It has been removed.

The messages are leaking implementation details that the pushlog is
stored in a database. The wording has been changed to be more generic
and concise.

--HG--
extra : commitid : 5uOuy7Q3CBh
extra : rebase_source : 065adaa1620be48d3a0d7b0adc46d805e06268c2
extra : amend_source : 1dfb038579a31eb8eabe2ffe231d5aeb02cdd2f0