scripts: advertise CDN URLs first (bug 1185261); r=RyanVM
This should be a rubber stamp review. Now that the updated bundleclone
extension with intelligent SNI handling should be deployed to
automation, we should be able to advertise CDN URLs as the primary
entries. If this assumption is wrong, we may see SSL-related job
failures in automation during clone operations. The recourse will be to
revert this changeset and update bundleclone manifests to the previous
behavior of listing S3 URLs first.

--HG--
extra : commitid : Jta90sINTCg
extra : amend_source : 52fcd67c369732e23df6289b051e964980307e26