Allow publishing from commandline regardless of whether or not the commit message has a reviewer (bug 1211564) r=gps
We allow publishing from the web ui without specifying a reviewer, at least
for some repositories, including gecko. This removes that requirement on the
command line.

This changed the behaviour of many of the tests which assume that we will not
publish commits without reviewers. To preserve the old behaviour a
config option is provided to disable automatic publishing when not desired.
This is set by default in helpers.sh.

--HG--
extra : rebase_source : 75317f8a882e467f1249db5e58f8e8b3f374fda6