autoland: store autoland request json directly (bug 1216947) r=glob
Currently we parse the request json blob into separate database fields when
the request is made. This implies a schema change every time we add a field to
the json. Since we only use the request data when actually transplanting the
changes, there's no benefit to having it in separate fields, and it adds
complexity when making changes to the request json schema.

This change will break querying for status for existing jobs in the autoland
database. Since the status endpoint is only used for debugging and all useful
information is mirrored back to MozReview this is acceptable. We'll have to
be careful when deploying that there are no pending autoland jobs which would
be lost when the database schema changes.

--HG--
extra : commitid : 9Hhy0r42Tc9
extra : rebase_source : a7c130cc84efc8ef6d90596f7bbd26dfc0b7139c
extra : amend_source : 1fb86041a632dd14abe502894b864f5c19912ddb