autoland: wait between retries if we hit an error posting to mozreview (bug 1217068) r=mdoglio
This adds a retry timeout if we hit an error posting to MozReview. The errors
we've seen so far are due to coding mistakes or misconfiguration so we don't
gain anything by retrying 4 times a second.

We also remove the requirement that the pingback_url be non-empty because it's
useful to leave the pingback_url empty when testing.

--HG--
extra : commitid : F2WKMEOeWOk
extra : amend_source : f5200d2cb2631dcdac34a9023177d75d21bce113