hgserver: re-enable clonebundles advertisement (bug 1221268)
Mercurial 3.6.1 is out and contains the fix for clone bundles that
broke `hg pull`.

This commit is a backout of aa86b7f9e406ea59082a3f2ae2f77d0e24abc1f5
with the addition of a slight wording change on the advertisement
message to note the bug in 3.6 and the requirement to run 3.6.1.

--HG--
extra : commitid : 6QNTolPmumo