autoland: ansibilize autoland provisioning and deployment (bug 1175529) r=gps
This adds ansible scripts to provision an autoland instance on aws as well as
to deploy a new version of autoland to an instance. The provisioning is
incomplete because some secrets are stored locally.

--HG--
extra : commitid : 6e8GRwVUEfs
extra : rebase_source : 63243ba4418ef855b1094c768b84d585fb5bc54f