mozreview: Add clear links to BMO comments when posting attachment (bug 1224631). r=dminor
I also changed the URL posted in the attachment to be to the diff instead of
the reviews, since that is more useful for reviewers in particular.
Finally, I removed the part about posting the changedesc, since that didn't
seem to work properly, and there's probably no point in duplicating it.

--HG--
extra : rebase_source : 61ff87b9834c9fbec70ded48462afe38dbdce227