mozreview: fix rendering of file diff review button (bug 1227508); r=mcote
To synchronise the rendering of the diff review buttons with the rendering
of the diff viewers we are using a client side function queue. This queue requires
every function to explictly declare when the next item should be executed. In bug 1227365
$.funcQueue('diff_files').next() was placed outside of the function, preventing the
the review button from rendering.

--HG--
extra : rebase_source : 056eda9ca4bbbd8d82b7b08610877baf669c260b