autoland: make rewritecommitdescriptions return base node for rebases (bug 1228007) r=gps
This changes rewritecommitdescriptions to return the base node for future
rebases. The base node is the oldest node present in commit descriptions or
the rewritten version of that node if it has a modified description. It is an
error if no nodes are present which match the specified commit descriptions.

--HG--
extra : rebase_source : 9c74994f03c72b28e2eab73d3e8d992626cc6300