vcsreplicator: add a context manager for performing kafka interactions
The intent is to catch Kafka errors and write them to a special log
and/or perform special actions when failures occur. In preparation for
this, we establish a generic context manager and call it throughout the
code. Features will be added in subsequent commits.

--HG--
extra : commitid : DkRAzN9ZDrN
extra : rebase_source : d6b3445a4d09a56db46efe2e51659d5bc1acca47