hgserver: don't require LDAP attribute update (bug 1232400); r=fubar
The LDAP master has been failing periodically for the last several days.
This was causing people to fail to SSH into hg.mozilla.org.

Since the LDAP master is only used for updating the last access
attribute and it isn't critical for this to succeed 100% of the time,
let's loosen the requirements around updating the last access attribute.

With this commit, we now ignore failures connecting to the LDAP master
server.

Tests demonstrating failure conditions have been added.

Failure to connect to the mirror continues to be fatal, but a test was
added. This is fatal because pash needs to validate that hg access is
present on the account.

We add tests demonstrating that failure to connect to the master can be
ignored.

--HG--
extra : rebase_source : cf8217a9e544f8315cff0ddb1d1d13d1658414e2