mozreview: update model when adding reviewers to trigger draft banner display (bug 1236999); r=smacleod
Review Board 2.5 added a check in the review request view's showBanner()
that this.model.get('hasDraft') is true. For reasons I don't fully
understand, the model isn't getting refreshed when we set a draft field
on it.

This patch sets the hasDraft attribute manually. I think ideally we'd
refresh the view's model instance, but I'm not sure how to do that.

--HG--
extra : commitid : F5MszvfzIPf