mozreview: add default reviewers the first time we add a diffset (bug 1239524); r=smacleod
The web UI enforces the requirement that a review request have
reviewers. This behavior is different from the web API, which doesn't
enforce this requirement.

As part of converting to the batch submission API, we failed to carry
over the setting of default reviewers from DiffResource.create() from
the Review Board Web API. This commit fixes that.

--HG--
extra : commitid : 9IuYOmKsaTT