mozreview: return proper error during batch submission processing (bug 1169664); r=smacleod
This function was returning a mix of a scalar value, a 2-tuple, or a
4-tuple. We performed a destructuring assignment assuming a 4-tuple.
This led to a ValueError and a HTTP 500.

We invent a new exception type to hold the return value for errors so we
return an appropriate error response instead of a HTTP 500.

--HG--
extra : commitid : 8ISnkTef4Nw
extra : rebase_source : f9c08ae5dbb7e78fd177d517d9848dfdf8116f42
extra : amend_source : 48fe7297ad987293126813a8c9c9cdc6c6626dcb