mozreview: refactor API key verification into standalone function (Bug 1244448). r=mcote r=gps
We're going to be verifying a Bugzilla API key in other resources so
pull out the code into a function we can call. It is a little bit
strange that we return either a user object or a WebAPIResponseError,
but we'll only be using this in API calls which will want to generate
those responses anyways.

--HG--
extra : commitid : 732g7tKkEFq
extra : rebase_source : 64f75eac39684b7029a3917a2f8c3ef79c7ba247