mozreview: copy draft_extra_data to extra_data for drafted CommitData (Bug 1244448). r=gps
We'll need a similar mechanism for having certain extra_data keys be moved from
draft_extra_data to extra_data as we used for the built-in extra_data. This is
particularly important because using a ReviewRequestField backed by our
CommitData won't take care of this for us - we'll need to do this for any key
which supports a draft value as well as the published value.

--HG--
extra : commitid : 1wUSJCTOgqf
extra : rebase_source : b21c9fedd1fe572bbb32502fead7352afeb7979d