testing: use ip address to set docker hostname if it is localunixsocket (bug 1245931) r=gps
If we get a hostname of "localunixsocket", we iterate over the docker
networks looking for the bridge interface so we can use the ip address to
set docker_hostname.

This only works if docker is listening on a tcp socket rather than a unix
domain socket. This was already the recommended configuration, I've updated
the documentation now that it is a requirement.

MozReview-Commit-ID: 2bR1Uk85Al3

--HG--
extra : rebase_source : 0aa563021ee77ff7638937a3f3e7bce7e6e6854d