vcsreplicator: refuse to operate on repos without a path mapping (bug 1240194); r=dminor
Currently, our mechanism for normalizing filesystem paths to wire paths
is optional. If no mapping is available, we use the filesystem path.

This caused issues with production when operating on a repository in a
path that wasn't supposed to be replicated. When the consumers
encountered the raw path, they blew up because they had no awareness of
the one-off path/repo.

This commit solves the issue by requiring a [replicationpathrewrites]
entry for a path. If we attempt to perform a replication event and we
can't resolve the repo's path via this config section, we abort.

We could have introduced a new section to declare which paths are under
replication control. But this feels like extra work for little gain. We
already have the [replicationpathrewrites] section, so it makes sense to
repurpose it and make it required. If nothing else, it forces people to
think about their replication configuration before blindly enabling the
extension.

MozReview-Commit-ID: 207Ib8k5izR

--HG--
extra : rebase_source : 274afb689b222d55ca6293932bea2e79d8c32768
extra : amend_source : d85cadcef8f9eb515e9142867549031e569340b5