testing: do not create autoland database in createdb.sql script
It appears that the postgres container will create the default database
for us automatically. Attempting to create it was causing the script to
error out and the container to stop. We can't just drop it if it exists
because it is not possible to drop the connected database.

--HG--
extra : commitid : DvIjt6zMskl
extra : rebase_source : c8cb34ee7660eab4c38f1fbaabf858443e88b8c3
extra : amend_source : 6eee243943892e727008c75b848091610274d0aa