autoland: set a max height on the commit confirmation dialog (bug 1247569) r=dminor
djblet's modal box doesn't restrict the height of content to the window,
resulting in a dialog with unreadable content when there are a lot of commits.
Unfortunately fixing this correctly by bounding the dialog to the window isn't
possible with RB css, so we just cap the maximum height at 600px.

MozReview-Commit-ID: HlTCNztV8ci

--HG--
extra : rebase_source : 54ee299ef502b1ab141eff7c6ed08c39894ac080