Bug 1247838 - Fix running git mozreview from a subdirectory. r=gps
Bug 1246998 broke it, and the push code in cinnabar assumes it's running
from the toplevel directory, which is a fine assumption in the
git-remote-hg helper, but not in the git cinnabar command. Until
cinnabar itself is fixed, work around it.

While here, add a test case for worktrees.

--HG--
extra : rebase_source : e34a28847a0d094c9089bad551476e5cd4b78929