mozreview: unify irc nick parsing and disallow periods (bug 1238353) r?smacleod
MozReview and mozautomation's commitparser disagree when it comes to deciding
if a period is valid in an IRC nick.  According to IRC a period is not valid,
which is mozautomation's behavour.  Updating mozreview and rbbz extensions to
use the RE defined in mozautomation instead of duplicating the code.

MozReview-Commit-ID: GlcmzbTUkIg

--HG--
extra : rebase_source : 406439ce08e7670ce9019dda748e3a99be8bf20a