hgmo: add a revset to return changesets relevant to automation (bug 1247802); r=dminor
Today, Firefox's automation isn't very intelligent about what jobs it
schedules for each push. It is overly aggressive and tends to schedule
everything all the time. This doesn't scale.

In order to schedule only what actually needs to be scheduled, you need
to know "what changed since last time." Build systems answer this
question by consulting the filesystem: if an mtime is newer than its
output file, the output file is stale and needs to be rebuilt. We can't
rely on this in automation because there are several build machines and
build machines tend to be frequently wiped. This means the state of the
mtimes on the filesystem can't be trusted and will yield incorrect
results for determining what changed.

Fortunately we have a versioned filesystem in the form of version
control. With version control, we can compute the set of files that
changed between 2 revisions very easily. The trick is identifying the 2
revisions.

Firefox automation schedules jobs from pushes to version control. When a
new push comes in, the revision at the head of that push is checked out
and built, etc. The head of the push is 1 revision. But we still need
an ancestor revision to compare it to so we can get the set of files
that changed.

We can't simply select the ancestor of the first revision in the push
because of the monolithic, many-headed Try repository. If you push 4
commits to Try then rewrite the tip-most commit and push to Try again,
that second push will only contain a single commit and selecting the
ancestor of that commit will only return 2 of the 4 commits relevant to
that Try push! A solution to this problem is to consult Mercurial's
phases information to supplement the changesets in a push with their non
public ancestors.

This commit introduces the "automationrelevant" revset. Given a
revision, it will select all revisions relevant to that revision for
purposes of running build automation. The revset is currently equivalent
to: nodesfrompush(REV) | (::REV & not public()). Although we don't
have a "nodesfrompush" or equivalent revset. We /could/ implement such a
revset. However, even if we had one, I'd rather automation not construct
their own revset because it is a bit more complicated and the behavior
of this revset may change over time. For example, I anticipate there
will eventually be an inbound-like non-publishing repo. For this repo,
"automationrelevant" should *not* extend with draft changesets. In the
future, we can teach the revset to behave differently when executed on
this type of repo.

After this lands, the plan is to teach automation to call a JSON API
that runs this revset and returns changeset metadata, from which it can
calculate the set of changed files and make intelligent scheduling
decisions.

MozReview-Commit-ID: BYYdV2m70cB

--HG--
extra : rebase_source : 89c8d30ed05d1368a0382c0893dd5ab8e7497fd9
extra : amend_source : 7744e5bab8e2c7417702c2b9d24e41204a049a9e