docker: better pinning of older setuptools and pip releases
Installing pip from get-pip.py would always download the latest
pip release. Sometime in the last few pip releases, the subsequent
downgrade to pip 6.0.8 later in the Dockerfile stopped working
because pip didn't like downgrading itself for whatever reason.

We change the image creation code to download archives of the
exact versions of setuptools and pip we'll use and install them
directly. Things appear to "just work."

We'll eventually want to upgrade to latest pip, which the latest
versions of peep now supports (which is the reason we're pinning
to such an old setuptools and pip). But that's for another commit.

MozReview-Commit-ID: FulTsBNaQy1

--HG--
extra : rebase_source : d609fec0563b3e5884e8d3b0a781049e6565a1e5