ansible/hg-ssh: add AssertPathExists on all units tied to master (bug 1268131)
We don't want to allow these units to start if the current machine is
not the master. Having the condition on hg-master.target wasn't enough:
we need these conditions on the individual units to prevent them from
starting on machines they shouldn't start on.

We use an assertion instead of a condition so attempts to start the
service manually result in a visible error. Conditions result in silent
non-service start.

--HG--
extra : amend_source : 928cb6a099164edd5f0e4b9a663afdfc295ab6f1