firefoxtree: update updated testedwith to drop old versions and add 3.8; r=ahal
Mercurial 3.8 is out. Our policy is to support at most the last 4
releases of Mercurial in our extensions. So update testedwith to reflect
that.

While we're here, we add the minimumhgversion variable, which is
recognized in Mercurial 3.7+ and aborts gracefully if the current
Mercurial version is older than specified.

MozReview-Commit-ID: Jvr6L4eRwmZ

--HG--
extra : rebase_source : ab9af329b8d51f2e239ba2b0b17172edc698dc18