ansible/docker-kafkabroker: wait for zookeeper before starting Kafka
I'm experiencing what appears to be race conditions when starting Kafka.
Kafka starts and we claim to create a topic, however the topic isn't
actually created.

Looking at the logs, it looks like the Kafka cluster isn't in a good
state when the topic creation request is sent. I reckon this is
because we start the Kafka cluster before Zookeeper is fully alive
and this results in a split brain of sorts.

This commit introduces some code into the Kafka startup process to
wait for Zookeeper to initialize before starting Kafka. This appears
to make the race condition disappear on my machine.

MozReview-Commit-ID: 3pk331d6Mn1

--HG--
extra : rebase_source : 57a0d3c5079716d07d4bd034cc63a4762b79f41f