Bug 1277103 - Always use git rev-parse to get git_dir. r=gps
When `git mozreview credentials` is executed by git/cinnabar from a
worktree, GIT_DIR is set to the directory of the private worktree
repository, and GIT_COMMON_DIR is not set, so we don't actually get the
right directory, and the credentials command fails with the error saying
running git mozreview configure is necessary.

So just always rely on git rev-parse.