generate-hg-s3-bundles: handle generaldelta repos
Previously, bundle generation assumed repos weren't generaldelta. We
hard-coded and used the "v1" bundle types. If these are used with
a generaldelta repo, Mercurial will re-encode things to previous delta,
which will take a lot of CPU time.

This commit teachs the bundle generation to handle generaldelta repos.
We now have a per-bundle type flag indicating whether it should be
used with generaldelta. The set of bundle types for generaldelta
vs non-generaldelta repos is currently distinct.