configwizard: backport util.versiontuple (bug 1281920); r=glob
util.versiontuple was added in Mercurial 3.7. The lack of it was
causing test failures when executed on hg < 3.7. Backport it if
necessary so the config wizard runs on older hg versions.

With this change, tests appear to pass on 3.5 and 3.6 with only
cosmetic differences (whitespace differences after prompt text
for reasons unknown to me).

MozReview-Commit-ID: FnzZ554Ut1X

--HG--
extra : rebase_source : 8765a5f24175127e742929d78fbcc5f37d875436