configwizard: work around ui.__class__ not working on hg <1.7 (bug 1282293); r=glob
See the inline comment for why we do this.

This enables the extension to load on hg 1.4, which is what RHEL 6
ships and is hopefully the oldest version we'll encounter in the wild.
It may work on earlier versions but I haven't tested.

MozReview-Commit-ID: 3kL844ZnNdp

--HG--
extra : rebase_source : 3d02403b68d20fbd4d5907e35ac3734d7f568e29