mozreview: hide trophies (bug 1267544) r=gps
The fish trophy is, honestly, a waste of time, and we'll never hit milestone
trophies.

Unregistering the trophie classes from an extension doesn't appear to work, and
if we modify the trophy code we'd have to update extra_data on all review
requests to force a recalculation.  Easiest option is to hide trophies with
css.

MozReview-Commit-ID: IIaSNP8uVZL

--HG--
extra : rebase_source : 80693edf81df93994b065e627c4d33d2dc9fdea9