vcsreplicator: have replication user operate on unfiltered repos
As part of adding more robust tests around replication of obsolete
changesets, I discovered a bug in the replication system.

If a mirror is out of sync and a subsequent push obsoletes a changeset,
the mirror may attempt to `hg pull -r <rev>` an obsolete changeset.
This will result in an error because the revision is filtered on the
server.

We want replication mirrors to stay fully in sync with the master.

This patch enables the hg server on the master node to operate on
the "unfiltered" repo - this is a view of the repo with all
changesets exposed. Typically Mercurial only exposes non-hidden
and non-secret changesets.

This causes no regressions in tests, so I'm pretty confident this
will just work.

MozReview-Commit-ID: 3K2RsozBlyl

--HG--
extra : rebase_source : e5d2aa11d889166d00b6600d68936296d8037d6b