testing: change start-local-mozreview to an executable script (bug 1286595) r=smacleod
start-local-mozreview is odd because it's presented as both a file that should
be sourced (according to the comments in the file), as well as a shell script
that should be executed (it's executable, and has a #! line).

Remove the ability to source the file; and update ./mozreview's shellinit
command to set MOZREVIEW_HOME to the path initialised by start-local-mozreview.

MozReview-Commit-ID: CJP0MQlbnlS

--HG--
extra : rebase_source : 96d5bff2122c5c547fc8d59126b47dd46bc003d7
extra : amend_source : 55ec5d1449a8ee51a3ea86c99870418edebdd0c7