mozreview: Mirror comments on parents to just the bug of the first child (bug 1226080); r=glob
This should have no real effect now because all commits in a single series are
currently associated with a single bug.  Regardless, we need to choose only
a single bug so that we can later associate the comment with a single
attachment.

If we later do allow multiple bugs in a single series, we'll have to pick
one attachment for each unique bug.  There is a TODO to that effect.

MozReview-Commit-ID: 1MkfDlhJMuR

--HG--
extra : rebase_source : a726ba5ac581e8555afa06f27665c6de23ad2752