mozreview: display last reviewed diff revision (Bug 1244884). r=glob
We will display the diff revision that was published most recently, but
before, the reviewers last review. Using these timestamps has problems
when drafts are left open and new revisions are added, but it should
cover the majority of cases for now.

In the future we'll want to improve upon the method used to calculate
the revision (see the TODO comment in the code) and possibly the
messaging / UI.

MozReview-Commit-ID: KsxFuls15By

--HG--
extra : rebase_source : 0a4031449be2a3e916201876edab718dd4b6a216