autoland: add more logging around transplant (Bug 1294729). r=glob
In order to diagnose errors with transplanting add more logging. This
is a bit of a temporary fix and autoland logging should have a bit of
a rework, but that will be left to a followup. This should be good
enough for now.

MozReview-Commit-ID: 3XlUG0Wfcc5

--HG--
extra : rebase_source : 30c34fe3db10f14919348b6f5ba2093f256ac141