vcsreplicator: handle unknown revisions when querying obsolescence metadata
Before, we assumed that output from `hg log -r <sha-1>` was empty if a
revision wasn't known. This is not true. In reality, Mercurial aborts
with an "unknown revision" error.

We change the code to reflect this. We also add a test case to catch it.

--HG--
extra : amend_source : aceb49c29de398a62fe9091fb68d9e3d6d7b139c