docker: disable InnoDB async I/O due to potential capacity issues (bug 1299337); r=glob
When attempting to run MozReview tests under -j6 on my machine, I was
seeing a number of intermittent test skips. Upon further investigation,
the bmoweb container was stopping while we were waiting for it to
start responding to HTTP requests. Looking at the logs of the container,
the underlying problem was MySQL was failing to start. The logs indicated
the InnoDB storage system was failing to come online. Specifically, the
aio component of it was having problems.

It appears there are global limits to the number of aio operations that
can be performed. My suspicion is that starting multiple MySQL servers
simultaneously hit such a limit.

Since aio isn't critical for our test needs, disabling it seems
acceptable.

This commit appears to reduce the intermittent failure rate on my
machine. Or, at least I'm no longer seeing MySQL initialization
failures (meaning the source of intermittent failures is now further
down the line).

MozReview-Commit-ID: BiXvvhtjXdz

--HG--
extra : rebase_source : aa424756739271962fdfcc2cba6b7790d308961d