ansible: run hg-reviewboard from CentOS 7 (bug 1299550); r=fubar,glob
hg.mozilla.org is 100% CentOS 7. We wish to transition our automation
away from CentOS 6 so it is consistent and more secure. (Notably, we're
running an old OpenSSH server on CentOS 6 and upgrading is difficult.)

This commit transitions the hg-reviewboard Docker environment and
ansible role to CentOS 7. The changes are relatively straightforward.

The httpd config needed porting to version 2.4.

We had to stop using paramiko in cluster startup because it doesn't
support ED25519 keys. So, we just write the SSH known_hosts file
directly. That should just work.

The deployment playbook has been updated to deploy to reviewboard-hg2.
This machine is available but not configured in the load balancer. We
need to remove deploying to reviewboard-hg1 because the ansible foo
no longer works on CentOS 6.

The only test fallout appears to be related to pushing over HTTP.
First, Python's built-in cgi.py module appears to enforce the
requirement that a Content-Type header be defined. We weren't sending
this before and this resulted in HTTP 500 errors in authenticated
requests. Second, httpd's error messages changed slightly.

MozReview-Commit-ID: 4FIGHimLGBr

--HG--
extra : rebase_source : d871f88803f0417c57aa749735b65dbe71519753