mozreview: "vendor" trychooser (bug 1304320) r=smacleod
Until trychooser is delivered over https (bug 1302631), "vendor" it from
https://hg.mozilla.org/build/tools/file/tip/trychooser

trychooser is slightly modified to remove jquery-ui.css and the reference to it
in index.html.  None of the styles it provides are actually used, and it was
preventing the mozreview egg from being built due to non-existent images that
are referenced.

MozReview-Commit-ID: IXTmtjBgpW4

--HG--
extra : rebase_source : 71974e19efe305188219ad5cc8f485fefdc32859