bug 1296641 - import new ansible playbooks and roles for autoland; r=fubar,smacleod
Apologies for this being a monolithic import.  Future work here will
be smaller and frequent.

This is a dump of the current state on the new ansible playbooks for
moving autoland to the new platform ops aws account.  This is by no
means in a completed state.  Any TODOs will be called out in bug 1262061

MozReview-Commit-ID: JGuxmyBM3Qf

--HG--
extra : rebase_source : d0ffa624feeb9f63b514830d86299aa9b8dca745