vcsreplicator: teach pulse notifier to skip messages (bug 1312238); r=fubar
We ran into an issue over the weekend where pulsenotifier.service got
wedged processing a bad message. This is the first instance of this in
months.

Unfortunately, we didn't have a good way for the pulse notification
daemon to skip over "bad" messages and I had to hack source code on
the server to restore service.

This commit introduces a "--skip" flag on the daemon that when
specified will instruct the daemon to skip the next message and
exit.

The operational docs for this service have been updated to reflect
how to invoke it in production.

MozReview-Commit-ID: GtKjsj4Lpb9

--HG--
extra : rebase_source : 84cf37701cb5fbce69dd9907f7490d92d78b64f2