ansible/docker-hg-ssh: run a mock S3 and SNS server (bug 1316952); r=glob
We're about to support sending hg.mo events to S3 and SNS. In order to
test this properly, we'll run mock servers that verify the publishing
code works as expected.

The "moto" package provides actual HTTP server implementations of
various AWS services. The implementations aren't very robust, of
course, but they are adequate for testing.

This commit adds a supervisor config to run local S3 and SNS HTTP
servers inside the hgssh Docker container.

We also introduce an "sns-subscription-server" script to run another
local HTTP server that will receive SNS "pings" in response to message
publications. This is also hooked up to supervisor. The server
simply logs its requests to a local file. This allows tests to cat the
file to see what happened. Ideally we'd ask the "moto" service for
previous events, but it doesn't appear to expose them for SNS.

The "configure-events-servers" helper script is introduced to bootstrap
the environment. It creates our future S3 bucket and SNS topic and
creates a subscription to the locally running SNS server.

Users of this infrastructure will be introduced in a subsequent
commit.

MozReview-Commit-ID: 91T50kaiSjc

--HG--
extra : rebase_source : b1ea273747d9148868e7a9ca043875e299b1044d