mozreview: the moved-from/to markers shouldn't indent code (bug 1319074); r=davidwalsh
Outdent moved-from and moved-to code, to accommodate the border marking the
range of lines impacted.

This required shrinking the marking border from 4px to 2px, to ensure there's
enough room in the margin for the border and a gap before the first character.

MozReview-Commit-ID: L6OLrnYzZCE

--HG--
extra : rebase_source : c082c46539c5a586d16203c767452453b12afe7f