mozautomation: handle empty reviewer string (bug 1321344); r=smacleod
97701ae2f512 (bug 1320251) changed the reviewer regular expression
slightly to allow optional characters after the review specifier so
we could detect bare "r?" or "r=". Unfortunately, parse_reviewers()
assumed this regexp match had a value and was subsequently failing
to operate on a None value.

This commit adds an explicit value check to stop that error.

The added tests triggered the same stack observed in the bug before
parse_reviewers() was updated. So I'm pretty confident this will
fix the issue.

MozReview-Commit-ID: BXZTMYLOewH