firefoxtree: handle revision 0 being hidden (bug 1325857); r=smacleod
The previous code would raise a FilteredRepoLookupError if revision 0
in the repo was hidden. This could happen if a user has obsolescence
enabled and rewrote the first commit in a repo.

I'm too lazy to write a test because it is an esoteric edge case.

MozReview-Commit-ID: GR0WIFeHnNb

--HG--
extra : rebase_source : ff3a3e5e8972950cc2ecebd9f9524c394c8855de